+++ rust/src/api/ffsend.rs
@@ -3,7 +3,7 @@ pub use ipp::model::JobState;
 
 use crate::{frb_generated::StreamSink, utils::ffsend_client::{self, FfSendTransferProgress}};
 
-pub fn ffsend_upload_file(host_url: String, file_path: String, download_filename: Option<String>, max_downloads: Option<u8>, expires_after_seconds: Option<usize>, update_sink: StreamSink<FfSendTransferProgress>) {
+pub fn ffsend_upload_file(host_url: String, file_path: String, download_filename: Option<String>, max_downloads: Option<u8>, expires_after_seconds: Option<u32>, update_sink: StreamSink<FfSendTransferProgress>) {
     ffsend_client::upload_file(host_url, file_path, download_filename, max_downloads, expires_after_seconds, update_sink)
 }
