+++ rust/src/models/images.rs
@@ -5,12 +5,12 @@ use num_derive::FromPrimitive;
 pub struct RawImage {
     pub format: RawImageFormat,
     pub data: Vec<u8>,
-    pub width: usize,
-    pub height: usize,
+    pub width: u32,
+    pub height: u32,
 }
 
 impl RawImage {
-    pub(crate) fn new_from_rgba_data(data: Vec<u8>, width: usize, height: usize) -> RawImage {
+    pub(crate) fn new_from_rgba_data(data: Vec<u8>, width: u32, height: u32) -> RawImage {
         Self {
             format: RawImageFormat::Rgba,
             data,
